Use a 32-bit notification flag

+++ src/Interprocess.Tests/CircularBufferTests.cs
@@ -113,7 +113,7 @@ public unsafe class CircularBufferTests
                 ReadOffset = 1,
                 WriteOffset = 2,
                 ReadLockTimestamp = long.MaxValue,
-                NotificationPending = long.MinValue
+                NotificationPending = int.MinValue
             };
             buffer.Write(value, 0);
             value.Should().BeEquivalentTo(*(QueueHeader*)ptr);
